+++ urbanairship-automation/src/main/java/com/urbanairship/automation/engine/AutomationEngine.kt
@@ -413,7 +413,7 @@ internal class AutomationEngine(
         }
 
         if (!preprocessDelay(data)) {
-            UALog.v {"Preprocessing delay interrupted, retrying" }
+            UALog.v {"Preprocessing delay interrupted $data, retrying" }
             processTriggeredSchedule(scheduleId)
         }
 
@@ -538,6 +538,7 @@ internal class AutomationEngine(
 
         val updated = updateState(data.schedule.identifier) {
             if (!it.isInState(listOf(AutomationScheduleState.TRIGGERED))) {
+                UALog.v { "Schedule $data no longer triggered" }
                 return@updateState it
             }